% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolateFunction.R
\name{interpolate}
\alias{interpolate}
\title{Data interpolation across soil depths}
\usage{
interpolate(Dataset, interCol, datRange)
}
\arguments{
\item{Dataset}{a dataframe of environmental observations}

\item{interCol}{the column over which the approx function embedded in interpolate() performs interpolation.}

\item{datRange}{The range of points to which the interCol column will be expanded.}
}
\value{
A dataframe of depth-interpolated values
}
\description{
Interpolate is a function that takes a dataframe with a limited number of observations and performs linear interpolation between the datapoints to generate an expanded dataset.The function takes a dataframe of observations and returns a dataframe containing all extrapolated values.
}
\examples{
interpolate(CO2, 4, 95:1000)
}
