% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageName.R
\docType{package}
\name{SoilMatrix}
\alias{SoilMatrix}
\alias{SoilMatrix-package}
\title{SoilMatrix: A package to organize functions needed for soil biogeochemical analyses.
SoilMatrix contains functions:
interpolate(), SoilCorMat(), SoilRainbowStats()}
\description{
SoilMatrix: A package to organize functions needed for soil biogeochemical analyses.
SoilMatrix contains functions:
interpolate(), SoilCorMat(), SoilRainbowStats()
}
\section{SoilMatrix functions}{

functions ...
}

