% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilDataGraphing_Functions.R
\name{SoilRainbowStats}
\alias{SoilRainbowStats}
\title{Color-coded scatter plots of soil dataset correlations}
\usage{
SoilRainbowStats(SoilDatSet, depthCol, numDepths, startCol, endCol)
}
\arguments{
\item{SoilDatSet}{A dataframe containing columns of numeric variables measured in soil samples}

\item{depthCol}{The column number in the dataframe for the column containing soil depth data.}

\item{numDepths}{The number of depths in/length of the soil depth column}

\item{startCol}{The column for which the user would like to start producing correlations. If the user wants to correlate all variables in the dataframe to each other, this value equals 1.}

\item{endCol}{The column for which the user would like to stop calculating correlations between variables. If the user wishes to correlate all variables, this number equals ncol(SoilDatSet)}
}
\value{
A matrix of scatter plots with corresponding correlation coefficients. Scattered points are color coordinated such that a rainbow-ordered plot indicates depth-dependence of the relationship. Lack of color order indicates no effect of depth on the relationship between 2 variables.
}
\description{
\code{SoilRainbowStats()} is a function that takes a dataframe of soil observations and analyzes the correlations between each column of variables. It then plots the correlation scatter plots with regression coefficients, and color codes plotted points in rainbow order according to soil depth. Shallower points are red and deeper points are violet.
}
\section{Note}{
 this function requires the colorspace package
}

\examples{
SoilRainbowStats(airquality, 6, 153, 1, 3 )
}
